+++ TicketsFinder_API/Services/IServices/INotificationService.cs
@@ -7,7 +7,7 @@ namespace TicketsFinder_API.Services.IServices
         public Task<int> CreateNotification(NotificationDTO notificationDTO);
         public Task<int> DeleteNotification(Guid notificationId);
         public Task<int> UpdateNotification(NotificationDTO notificationDTO);
-        public Task<IEnumerable<NotificationDTO>> GetNotifications(int chatId);
-        public Task<int> CheckCount(int chatId);
+        public Task<IEnumerable<NotificationDTO>> GetNotifications(long chatId);
+        public Task<int> CheckCount(long chatId);
     }
 }
